+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.6.4 (...)
 - convert spaces in `_` when downloading an OpenVPN configuration through the
   portal, fixes import in NetworkManager (Linux) (#92)
+- return `reason` through API when checking validity of certificate
 
 ## 1.6.3 (2018-04-20)
 - fix `/check_certificate` response to match API (#32)